bug fix: Search bar functionality break
After breaking the if/else statement into a switch case statement,
a bug was introduced in which the search results for a state and no city
was coming back from our cleaner functions empty or with very little data.

The switch case statement has been revised to reflect the nature of the
comparison being made of the switch expression against the case expression.

Additionally, I have corrected an invalid comparison in an if statement
nested within the case block.

This commit returns the search functionality to its expected behavior.